Cleanup files and prepare for Pull Request.

+++ docker-compose/postgresql/initdb.sh
@@ -1,21 +1,25 @@
 #!/bin/bash
 
-# This file is skipped automatically when 
-#     - in docker compose there is an image already defined or 
-#     - the database is already installed in the container.
+# This file will be automatically executed only when there is no database installed.
+# If a database is detected, this file will be skipped, whatever the contents of the database are.
 
 echo "Starting DB initialization."
-echo "Check if database 'adempiere' exists."
 
-#if [ "$( psql -U $POSTGRES_USER -tAc "SELECT 1 FROM pg_roles WHERE rolname='adempiere'" )" != '1' ]
-#if [[ -z `psql -Atqc '\list adempiere' postgres` ]]  # Test database existence
-#then
-    echo "The database 'adempiere' does not exist -->> it will be created and restored"
+echo "Check if user 'adempiere' exists."
+if [ "$( psql -U $POSTGRES_USER -tAc "SELECT 1 FROM pg_roles WHERE rolname='adempiere'" )" != '1' ]
+then
+    echo "The role 'adempiere' does not exist -->> it will be created and restored"
     createuser -U postgres adempiere -dlrs
     psql -U postgres -tAc "alter user adempiere password 'adempiere';"
-    createdb -U adempiere adempiere
+fi
 
+echo "Check if database 'adempiere' exists."
+if [[ -z `psql -Atqc '\list adempiere' postgres` ]]  # Test database existence
+then
+    echo "The database 'adempiere' does not exist -->> it will be created and restored"
     echo "Restore of database 'adempiere' starting..."
+    createdb -U adempiere adempiere
+    
     echo "Check if a seed restore file exists ($POSTGRES_DEFAULT_RESTORE_FILE)"
     if [ -f "$POSTGRES_DEFAULT_RESTORE_FILE" ]
     then
@@ -29,31 +33,39 @@ echo "Check if database 'adempiere' exists."
       echo "Downloading ADempiere artifact from Github.... It may take some time"
       wget --no-check-certificate --content-disposition $ADEMPIERE_GITHUB_ARTIFACT
       echo "Result from ls -la: " `ls -la`
-      echo "Unpack $ADEMPIERE_GITHUB_COMPRESSED_FILE here.... It may take some time"
-      tar -xvf $ADEMPIERE_GITHUB_COMPRESSED_FILE Adempiere/data/Adempiere_pg.dmp  -C .
-      echo "Result from ls -la: " `ls -la`
-      echo "Rename Adempiere_pg.dmp to $POSTGRES_RESTORE_FILE_NAME. Any existing file with same name will disappear!"
-      mv Adempiere/data/Adempiere_pg.dmp $POSTGRES_RESTORE_FILE_NAME
-      echo "Result from ls -la: " `ls -la`
-      rm -rf Adempiere
-      echo "Delete $ADEMPIERE_GITHUB_COMPRESSED_FILE    "
-      rm $ADEMPIERE_GITHUB_COMPRESSED_FILE
-      echo "Result from ls -la: " `ls -la`
+      
+      echo "Check file $ADEMPIERE_GITHUB_COMPRESSED_FILE was downloaded"
+      if [ -f "$ADEMPIERE_GITHUB_COMPRESSED_FILE" ]
+      then
+      echo "File $ADEMPIERE_GITHUB_COMPRESSED_FILE was downloaded"
+        echo "Unpack $ADEMPIERE_GITHUB_COMPRESSED_FILE here.... It may take some time"
+        tar -xvf $ADEMPIERE_GITHUB_COMPRESSED_FILE Adempiere/data/Adempiere_pg.dmp  -C .
+        echo "Result from ls -la: " `ls -la`
+        echo "Rename Adempiere_pg.dmp to $POSTGRES_RESTORE_FILE_NAME. Any existing file with same name will disappear!"
+        mv Adempiere/data/Adempiere_pg.dmp $POSTGRES_RESTORE_FILE_NAME
+        echo "Result from ls -la: " `ls -la`
+        rm -rf Adempiere
+        echo "Delete $ADEMPIERE_GITHUB_COMPRESSED_FILE"
+        rm $ADEMPIERE_GITHUB_COMPRESSED_FILE
+        echo "Result from ls -la: " `ls -la`
+      else
+        echo "ERROR: File $ADEMPIERE_GITHUB_COMPRESSED_FILE could not be downloaded."
+        echo "Process will be stopped."
+        exit 1
+      fi
     fi
 
-    # The following commands are not used anymore. Left because od legacy.
-    #psql -U adempiere -d adempiere < Adempiere/data/Adempiere_pg.dmp
-    #pg_restore -U adempiere -d adempiere < /tmp/seed.backup -v  # In case Backup was created with pg_dump
     echo "Ready to start DB restore"
     echo "Restoring  ADempiere artifact from Github.... It may take some time"
+    # The following command is not used anymore. Left because of legacy.
+    #pg_restore -U adempiere -d adempiere < /tmp/seed.backup -v  # In case Backup was created with pg_dump
     psql -U adempiere -d adempiere < $POSTGRES_DEFAULT_RESTORE_FILE
-    #psql -U adempiere -d adempiere < Adempiere_pg.dmp
     echo "Restore of database 'adempiere' finished"
     echo "Test that database 'adempiere' was created:"
     psql -Atqc '\list adempiere'
-#else
-#    echo "Database 'adempiere' does already exist -->> it needs not be created"
-#fi
+else
+    echo "Database 'adempiere' does already exist -->> it needs not be created"
+fi
 
 AFTER_RUN_DIR="/tmp/after_run"
 for file in $AFTER_RUN_DIR/*; do
